Fb  secure api (#13)
* small fixes + DTO added

* added user repo

* fixed null possible references

* login controller

* added empty auth service

* removed warnings

* renaming

* DB creation separation

* Find user by username

* .

* Hash generator

* Working debug JWT generation

* added JWT as api security protocol

* small fixes + DTO added

* added user repo

* fixed null possible references

* login controller

* added empty auth service

* removed warnings

* renaming

* DB creation separation

* Find user by username

* .

* Hash generator

* Working debug JWT generation

* added JWT as api security protocol

* merge duplicate

* ignore db cache

* housekeeping

* small fixes + DTO added

* added user repo

* fixed null possible references

* login controller

* added empty auth service

* removed warnings

* renaming

* DB creation separation

* Find user by username

* .

* Hash generator

* Working debug JWT generation

* added JWT as api security protocol

* small fixes + DTO added

* added user repo

* fixed null possible references

* login controller

* added empty auth service

* renaming

* DB creation separation

* Find user by username

* Hash generator

* Working debug JWT generation

* added JWT as api security protocol

* ignore db cache

* housekeeping

* Update GroceryListController.cs

* Update GroceryListDTO.cs

* Update GroceryList.cs

* Update GroceryList.cs

+++ Backend/Infrastructure/GroceryListRepository.cs
@@ -36,6 +36,12 @@ public class GroceryListRepository : IRepository<GroceryList>
     public GroceryList Delete(int id)
     {
         var groceryList = _dbContext.GroceryListsTable.Find(id);
+
+        if (groceryList == null)
+        {
+            throw new NullReferenceException("Grocery list not found");
+        }
+
         _dbContext.GroceryListsTable.Remove(groceryList);
         _dbContext.SaveChanges();
